--- COMPILING_linux
@@ -1,19 +1,0 @@
-Compiling instructions for Linux
-OpenTitus 0.8.0
-
-OpenTitus is released under the Gnu GPL version 3 or (at your option) any later version of the license, and it is released "as is"; without any warranty.
-
-
-Compile time dependencies:
-LibSDL development headers
-LibSDLMixer library
-Chocolate-Doom (1.6.0)'s OPL library (opl/libopl_linux.a), check out opl/opl_readme.txt for further details
-
-To compile, use the following commands:
-
-All:
-make -f Makefile_linux
-
-
-Clean:
-make -f Makefile_linux clean
