+++ bin/titus.conf
@@ -4,7 +4,7 @@
 reswidth 320
 resheight 200
 bitdepth 32
-devmode 1
+devmode 0
 ingamewidth 320
 ingameheight 192
