--- builds/README
@@ -1,11 +1,0 @@
-This folder contains precompiled versions of OpenTitus. Move them to the parent folder to make use of them.
-
-Versions:
-
-1.0.0 beta with adlib audio:
-Win32
-Linux
-Dingux (buggy)
-
-1.0.0 beta without adlib audio:
-Dingux
