+++ src/audio.c
@@ -13,11 +13,7 @@
  */
 
 //Probably not the best way, but it works...
-#define HAVE_CONFIG_H 1
-
-#ifdef HAVE_CONFIG_H
 #include "config.h"
-#endif
 
 #ifdef AUDIO_ENABLED
