+++ src/backbuffer.c
@@ -1,0 +1,4 @@
+#include "backbuffer.h"
+
+SDL_Surface *screen; //Backbuffer
+SDL_Surface *tilescreen; //Tile screen
