+++ src/config.h
@@ -31,7 +31,7 @@
 
 /* Define to 1 if you want to compile with debug support */
 //#undef DEBUG_VERSION
-#define DEBUG_VERSION 1
+//#define DEBUG_VERSION 1
 
 /* Define to 1 if you want to compile with audio support */
 //#undef AUDIO_ENABLED
