+++ src/gates.h
