+++ src/menu.c
@@ -39,12 +39,7 @@
 #include "globals.h"
 #include "common.h"
 
-//Probably not the best way, but it works...
-#define HAVE_CONFIG_H 1
-
-#ifdef HAVE_CONFIG_H
 #include "config.h"
-#endif
 
 #ifdef AUDIO_ENABLED
 #include "audio.h"
