+++ src/original.c
@@ -34,6 +34,35 @@
 #include "settings.h"
 #include "definitions.h"
 
+SDL_Color orig_palette_colour[16];
+SDL_Color orig_palette_level_colour[16];
+SDL_Color orig_palette_font_colour[16];
+
+uint8 spritewidth[SPRITECOUNT];
+uint8 spriteheight[SPRITECOUNT];
+uint8 spritecollwidth[SPRITECOUNT];
+uint8 spritecollheight[SPRITECOUNT];
+uint8 spriterefwidth[SPRITECOUNT];
+uint8 spriterefheight[SPRITECOUNT];
+
+int16 anim_player[ANIM_PLAYER_COUNT][ANIM_PLAYER_MAX];
+int16 anim_enemy[NMI_ANIM_TABLE_COUNT];
+uint8 NMI_POWER[ORIG_LEVEL_COUNT];
+uint8 LEVEL_MUSIC[ORIG_LEVEL_COUNT];
+
+int16 anim_zoubida[ORIG_ANIM_MAX];
+int16 anim_moktar[ORIG_ANIM_MAX];
+int16 anim_smoke[ORIG_ANIM_MAX];
+int16 COEUR_POS[ORIG_ANIM_MAX * 2];
+
+uint8 object_maxspeedY[ORIG_OBJECT_COUNT];
+bool object_support[ORIG_OBJECT_COUNT]; //not support/support
+bool object_bounce[ORIG_OBJECT_COUNT]; //not bounce/bounce against floor + player bounces (ball, all spring, yellow stone, squeezed ball, skateboard)
+bool object_gravity[ORIG_OBJECT_COUNT]; //no gravity on throw/gravity (ball, all carpet, trolley, squeezed ball, garbage, grey stone, scooter, yellow bricks between the statues, skateboard, cage)
+bool object_droptobottom[ORIG_OBJECT_COUNT]; //on drop, lands on ground/continue below ground(cave spikes, rolling rock, ambolt, safe, dead man with helicopter)
+bool object_no_damage[ORIG_OBJECT_COUNT]; //weapon/not weapon(cage)
+
+
 int initoriginal() {
     int i, j;
