+++ src/settings.c
@@ -31,12 +31,42 @@
 #include <string.h>
 #include "settings.h"
 
-//Probably not the best way, but it works...
-#define HAVE_CONFIG_H 1
-
-#ifdef HAVE_CONFIG_H
 #include "config.h"
-#endif
+
+char spritefile[256];
+char levelfiles[16][256]; //16 levels in moktar, 15 levels in titus
+char tituslogofile[256];
+int tituslogoformat;
+char titusintrofile[256];
+int titusintroformat;
+char titusmenufile[256];
+int titusmenuformat;
+char titusfinishfile[256];
+int titusfinishformat;
+char fontfile[256];
+int levelcount;
+int devmode;
+int reswidth;
+int resheight;
+int bitdepth;
+int ingamewidth;
+int ingameheight;
+int videomode;
+int game;
+
+char levelcode[16][5];
+char leveltitle[16][41];
+
+char moduleintrofile[256]; //.mod file
+int  moduleintrofileloop; //loop info
+char moduleprelevelfile[256];
+int  moduleprelevelfileloop;
+char modulelevelfile[6][256]; //6 different level files
+int  modulelevelfileloop[6];
+char modulegameoverfile[256];
+int  modulegameoverfileloop;
+char modulelevel[16]; //Link to modulelevelfiles
+
 
 int readconfig(char *configfile) {
     char line[300], tmp[256];
