+++ src/tituserror.c
@@ -1,0 +1,3 @@
+char lasterror[200];
+int lasterrornr; //Only to be used when needed, f.ex. when return value is not int (f.ex. in function SDL_Text) (maybe this always should be used?)
+void checkerror(void);
